Fixup PPU address mapping in mappers.  Mapper1 fairly functional now.

+++ SaltyLib/SaltyLib.Shared/NES/NESRom.h
@@ -61,7 +61,7 @@ public:
 
 	bool HasChrRom() const;
 	const byte* GetChrRom() const;
-	uint16_t CbChrRomData() const { return m_header.CbChrRomData(); }
+	uint32_t CbChrRomData() const { return m_header.CbChrRomData(); }
 
 	PPU::MirroringMode GetMirroringMode() const { return m_header.GetMirroringMode(); }
